Refactors AllteamInvitations
